AgregarAlCarrito funcionando, GetCarrito funcionando, se agregaron dr…

+++ Carrito-C/Carrito-C/Models/ComprasItem.cs
@@ -24,6 +24,6 @@ namespace Carrito_C.Models
 
 
         [DataType(DataType.Currency)]
-        public int Subtotal { get; set; }
+        public double Subtotal { get; set; }
     }
 }
